DevBG.Python.Meetup.May-2021/demo/requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
flake8==3.9.1
idna==2.10
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mccabe==0.6.1
paho-mqtt==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
requests==2.25.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
